Use pagination to retrieve full list of users
Previously, the script would only retrieve the first 25 users
(PagerDuty's default). We can increase this to 100 (PagerDuty's
maximum) but we need to use pagination, too, to fetch users beyond
the first 100 (we currently have 192). See:
https://developer.pagerduty.com/docs/ZG9jOjExMDI5NTU4-pagination#classic-pagination

For ease of test writing, and to better encapsulate concerns,
the `['users']` retrieval has been moved from `bin/schedule` to
the more appropriate home of the `PagerDuty` class.