Retrieve full list of schedules
We currently have 42 PagerDuty schedules, which is more items than
the default PagerDuty API limit of 25.

We could have reused the pagination logic from the `users` method,
but YAGNI. It seems unlikely we'll breach the 100 item limit where
pagination becomes necessary. The code and tests are much easier
to read without the hassle of pagination and the inevitable
abstraction layer we'd end up building to DRY up the shared logic.