Add rubocop gem and config
Here I have disable Style/DateTime as it wasn't a safe autocorrect. Rubocop wants to change to either Date or Time.
Date is not appropriate in this case, but Time might be. In this case Time and DateTime differ in the handling of DSTetc it's better not to change as part of a linting fix.
We should probably work out how to compare/test these differences at a later stage.