Remove poor bank holidays support from pay-pagerduty
We’re moving towards a model where pay-pagerduty will contain less logic: its only job should be to take a bunch of names and dates, and synchronise those with PagerDuty. The bank holidays logic (if any) should be applied earlier on, in govuk-rota-generator, as part of generating said list of names and dates.